# swagger.yaml
openapi: 3.0.0
paths:
  "/v1/health":
    get:
      parameters: []
      responses:
        '200':
          description: ''
  "/v1/asset/upload":
    put:
      parameters: []
      requestBody:
        required: true
        description: Asset files
        content:
          multipart/form-data:
            schema:
              "$ref": "#/components/schemas/FilesUploadDto"
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UploadResponseDto"
  "/v1/content/{userDsnpId}/broadcast":
    post:
      parameters:
      - name: userDsnpId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BroadcastDto"
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AnnouncementResponseDto"
  "/v1/content/{userDsnpId}/reply":
    post:
      parameters:
      - name: userDsnpId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReplyDto"
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AnnouncementResponseDto"
  "/v1/content/{userDsnpId}/reaction":
    post:
      parameters:
      - name: userDsnpId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReactionDto"
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AnnouncementResponseDto"
  "/v1/content/{userDsnpId}":
    put:
      parameters:
      - name: userDsnpId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateDto"
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AnnouncementResponseDto"
    delete:
      parameters:
      - name: userDsnpId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TombstoneDto"
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AnnouncementResponseDto"
  "/v1/profile/{userDsnpId}":
    put:
      parameters:
      - name: userDsnpId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProfileDto"
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AnnouncementResponseDto"
info:
  title: Content Publishing Service API
  description: Content Publishing Service API
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Enter JWT token
    cookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    FilesUploadDto:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
      required:
      - files
    UploadResponseDto:
      type: object
      properties:
        assetIds:
          type: array
          items:
            type: string
      required:
      - assetIds
    AssetReferenceDto:
      type: object
      properties:
        referenceId:
          type: string
          minLength: 1
        height:
          type: number
          minimum: 1
        width:
          type: number
          minimum: 1
        duration:
          type: string
          pattern: DURATION_REGEX
      required:
      - referenceId
    AssetDto:
      type: object
      properties:
        type:
          type: string
          enum:
          - link
          - image
          - audio
          - video
        references:
          type: array
          items:
            "$ref": "#/components/schemas/AssetReferenceDto"
        name:
          type: string
          minLength: 1
        href:
          type: string
          minLength: 1
      required:
      - type
    TagDto:
      type: object
      properties:
        type:
          type: string
          enum:
          - mention
          - hashtag
        name:
          type: string
          minLength: 1
        mentionedId:
          type: string
          minLength: 1
          pattern: DSNP_USER_URI_REGEX
      required:
      - type
    LocationDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        accuracy:
          type: number
          minimum: 0
          maximum: 100
        altitude:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
          minimum: 0
        units:
          type: string
          enum:
          - cm
          - m
          - km
          - inches
          - feet
          - miles
      required:
      - name
    NoteActivityDto:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        published:
          type: string
          pattern: ISO8601_REGEX
        assets:
          type: array
          items:
            "$ref": "#/components/schemas/AssetDto"
        name:
          type: string
        tag:
          type: array
          items:
            "$ref": "#/components/schemas/TagDto"
        location:
          "$ref": "#/components/schemas/LocationDto"
      required:
      - content
      - published
    BroadcastDto:
      type: object
      properties:
        content:
          "$ref": "#/components/schemas/NoteActivityDto"
      required:
      - content
    AnnouncementResponseDto:
      type: object
      properties:
        referenceId:
          type: string
      required:
      - referenceId
    ReplyDto:
      type: object
      properties:
        inReplyTo:
          type: string
          pattern: DSNP_CONTENT_URI_REGEX
        content:
          "$ref": "#/components/schemas/NoteActivityDto"
      required:
      - inReplyTo
      - content
    ReactionDto:
      type: object
      properties:
        emoji:
          type: string
          minLength: 1
          pattern: DSNP_EMOJI_REGEX
        apply:
          type: number
          minimum: 0
          maximum: 255
        inReplyTo:
          type: string
          pattern: DSNP_CONTENT_URI_REGEX
      required:
      - emoji
      - apply
      - inReplyTo
    UpdateDto:
      type: object
      properties:
        targetContentHash:
          type: string
          pattern: DSNP_CONTENT_HASH_REGEX
        targetAnnouncementType:
          type: string
          enum:
          - broadcast
          - reply
        content:
          "$ref": "#/components/schemas/NoteActivityDto"
      required:
      - targetContentHash
      - targetAnnouncementType
      - content
    TombstoneDto:
      type: object
      properties:
        targetContentHash:
          type: string
          pattern: DSNP_CONTENT_HASH_REGEX
        targetAnnouncementType:
          type: string
          enum:
          - broadcast
          - reply
      required:
      - targetContentHash
      - targetAnnouncementType
    ProfileActivityDto:
      type: object
      properties:
        icon:
          type: array
          items:
            "$ref": "#/components/schemas/AssetReferenceDto"
        summary:
          type: string
        published:
          type: string
          pattern: ISO8601_REGEX
        name:
          type: string
        tag:
          type: array
          items:
            "$ref": "#/components/schemas/TagDto"
        location:
          "$ref": "#/components/schemas/LocationDto"
    ProfileDto:
      type: object
      properties:
        profile:
          "$ref": "#/components/schemas/ProfileActivityDto"
      required:
      - profile